Develop

+++ ticketsystem-backend/src/main/java/com/kevdeto/ticketsystem/domain/dto/response/TicketResponseDTO.java
@@ -1,12 +1,13 @@
 package com.kevdeto.ticketsystem.domain.dto.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
 public record TicketResponseDTO(
 	    Long id,
 	    LocalDateTime issueDate,
-	    Double total,
+	    BigDecimal total,
 	    Long businessId,
 	    List<TicketItemResponseDTO> items
 	) {}
